+++ .prettierignore
@@ -54,10 +54,6 @@ Hola10
 # dotenv environment variable files
 .env*
 
-# gatsby files
-.cache/
-public
-
 # Mac files
 .DS_Store
