+++ content/yml/skills.yaml
@@ -24,3 +24,4 @@ frameworks_libraries:
 - Scikit-Learn
 - Ionic
 - Gatsby
+- NextJS
