+++ gatsby-browser.js
@@ -6,11 +6,3 @@
 
 // You can delete this file if you're not using it
 
-import "./src/assets/css/base.css";
-import "./src/assets/css/main.scss";
-import "./src/assets/css/nirmalhk7.scss";
-import "./src/assets/css/vendor.scss";
-
-import "./src/assets/css/bootstrap-grid.css";
-import "./src/assets/scss/custom.scss";
-import "./src/assets/scss/index.scss";
