+++ next.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+    eslint: {
+      // Warning: This allows production builds to successfully complete even if
+      // your project has ESLint errors.
+      ignoreDuringBuilds: true,
+    },
+  };
