+++ package.json
@@ -13,30 +13,12 @@
     "autoprefixer": "10.3.4",
     "cross-fetch": "3.1.4",
     "eslint-plugin-react": "7.25.1",
-    "gatsby": "3.13.0",
-    "gatsby-image": "3.11.0",
-    "gatsby-plugin-feed": "3.13.0",
-    "gatsby-plugin-google-gtag": "3.13.0",
-    "gatsby-plugin-manifest": "3.13.0",
-    "gatsby-plugin-nprogress": "3.13.0",
-    "gatsby-plugin-offline": "4.13.0",
-    "gatsby-plugin-pdf": "0.0.1",
-    "gatsby-plugin-postcss": "4.13.0",
-    "gatsby-plugin-purgecss": "6.0.2",
-    "gatsby-plugin-react-helmet": "4.13.0",
-    "gatsby-plugin-sass": "4.13.0",
-    "gatsby-plugin-sharp": "3.13.0",
-    "gatsby-remark-images": "5.10.0",
-    "gatsby-source-filesystem": "3.13.0",
-    "gatsby-transformer-remark": "4.10.0",
-    "gatsby-transformer-sharp": "3.13.0",
-    "gatsby-transformer-yaml": "3.13.0",
     "jquery": "3.6.0",
     "moment": "2.29.1",
-    "pdf2html": "2.5.2",
     "postcss": "^8.3.5",
     "prop-types": "15.7.2",
     "purgecss": "4.0.3",
+    "next": "latest",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-helmet": "6.1.0",
@@ -55,27 +37,19 @@
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-sonarjs": "0.10.0",
     "eslint-webpack-plugin": "3.0.1",
-    "gatsby-plugin-eslint": "4.0.0",
     "gh-pages": "3.2.0",
     "husky": "6.0.0",
     "lint-staged": "11.0.0",
     "prettier": "2.3.0"
   },
-  "keywords": [
-    "gatsby"
-  ],
   "license": "0BSD",
   "scripts": {
-    "build": "gatsby build",
-    "develop": "gatsby develop",
     "format": "prettier --write \"**/*.{js,jsx,ts,tsx,json,md}\"",
-    "start": "gatsby develop",
-    "start:draft": "GATSBY_DRAFT=true gatsby develop",
-    "serve": "gatsby serve",
     "packages:update": "npm update --save-dev && npm update --save",
-    "clean": "gatsby clean",
-    "lint": "eslint . --fix",
-    "test": "echo \"Write tests! -> https://gatsby.dev/unit-testing\" && exit 1"
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "lint": "eslint . --fix"
   },
   "repositgory": {
     "type": "git",
