+++ src/components/contactme.js
@@ -1,7 +1,7 @@
 import { faTimes } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import React from "react";
-import "./../../assets/css/contactme_module.scss";
+import "./../assets/css/contactme.module.scss";
 
 class ContactForm extends React.Component {
   constructor(props) {
