+++ src/components/link.js
@@ -1,0 +1,11 @@
+import Link from 'next/link';
+
+function XLink({ title, to, children }) {
+  return (
+    <Link href={to} title={title}>
+      {children}
+    </Link>
+  );
+}
+
+export default XLink;
