+++ src/components/navbar.js
@@ -1,7 +1,7 @@
 import React from "react";
-import { Link } from "gatsby";
 import Scrollspy from "react-scrollspy";
-import "./../../assets/css/navbar_module.scss";
+import "../assets/css/navbar.module.scss";
+import Link from "./link";
 
 const Navbar = () => {
   const navbarList = ["Home", "About", "Projects", "Blog", "Contact"];
