+++ src/components/quote.js
