+++ src/components/seo.js
@@ -1,88 +1,21 @@
-/**
- * SearchEnggOp component that queries for data with
- *  Gatsby's useStaticQuery React hook
- *
- * See: https://www.gatsbyjs.org/docs/use-static-query/
- */
+import Head from 'next/head';
+import config from '../config';
 
-import React from "react";
-import PropTypes from "prop-types";
-import { Helmet } from "react-helmet";
-import { useStaticQuery, graphql } from "gatsby";
-
-function SearchEnggOp({ description, lang, meta, title }) {
-  const { site } = useStaticQuery(
-    graphql`
-      query {
-        site {
-          siteMetadata {
-            title
-            description
-            author
-          }
-        }
-      }
-    `
-  );
-
-  const metaDescription = description || site.siteMetadata.description;
+export default function SEO({ description, title }) {
+  const siteTitle = config.title;
 
   return (
-    <Helmet
-      htmlAttributes={{
-        lang,
-      }}
-      meta={[
-        {
-          name: `description`,
-          content: metaDescription,
-        },
-        {
-          property: `og:title`,
-          content: title,
-        },
-        {
-          property: `og:description`,
-          content: metaDescription,
-        },
-        {
-          property: `og:type`,
-          content: `website`,
-        },
-        {
-          name: `twitter:card`,
-          content: `summary`,
-        },
-        {
-          name: `twitter:creator`,
-          content: site.siteMetadata.author,
-        },
-        {
-          name: `twitter:title`,
-          content: title,
-        },
-        {
-          name: `twitter:description`,
-          content: metaDescription,
-        },
-      ].concat(meta)}
-      title={title}
-      titleTemplate={`%s | ${site.siteMetadata.title}`}
-     />
+    <Head>
+      <title>{`${title} | ${siteTitle}`}</title>
+      <meta content={description} name="description" />
+      <meta content="website" property="og:type" />
+      <meta content={title} property="og:title" />
+      <meta content={description} property="og:description" />
+      <meta content={siteTitle} property="og:site_name" />
+      <meta content="summary" property="twitter:card" />
+      <meta content={config.social.twitter} property="twitter:creator" />
+      <meta content={title} property="twitter:title" />
+      <meta content={description} property="twitter:description" />
+    </Head>
   );
 }
-
-SearchEnggOp.defaultProps = {
-  lang: `en`,
-  meta: [],
-  description: ``,
-};
-
-SearchEnggOp.propTypes = {
-  description: PropTypes.string,
-  lang: PropTypes.string,
-  meta: PropTypes.arrayOf(PropTypes.object),
-  title: PropTypes.string.isRequired,
-};
-
-export default SearchEnggOp;
