+++ src/components/social.js
@@ -2,7 +2,7 @@ import { faAngellist, faDev, faGithub, faGoodreads, faLinkedin } from "@fortawes
 import { faRss } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import React from "react";
-import "../../assets/css/social_module.scss";
+import "../assets/css/social.module.scss";
 
 // TODO fix colors of social icons
 export const social = [
