+++ src/config.js
@@ -1,0 +1,11 @@
+export default {
+    title: 'Starter Blog',
+    author: {
+        name: 'Lee Robinson',
+        summary: 'who loves Next.js.',
+    },
+    description: 'A starter blog converting Gatsby -> Next.',
+    social: {
+        twitter: 'leeerob',
+    },
+}
