+++ src/pages/_app.js
@@ -1,0 +1,13 @@
+import "../assets/css/base.css";
+import "../assets/css/main.scss";
+import "../assets/css/nirmalhk7.scss";
+import "../assets/css/vendor.scss";
+
+import "../assets/css/bootstrap-grid.css";
+import "../assets/scss/custom.scss";
+import "../assets/scss/index.scss";
+
+// This default export is required in a new `pages/_app.js` file.
+export default function MyApp({ Component, pageProps }) {
+  return <Component {...pageProps} />;
+}
